second-approach/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.10
SQLAlchemy==2.0.37
gunicorn
flask-socketio
eventlet
Flask-Login==0.6.3
python-dotenv==1.0.0
Flask-Mail==0.9.1
redis==5.0.1
